creating add-image funtionality

+++ src/collections/application/use-cases/delete-collection.use-case copy.ts
@@ -1,4 +1,4 @@
-import { Injectable } from "@nestjs/common";
+import { Injectable, NotFoundException } from "@nestjs/common";
 import { CollectionRepository } from "src/collections/domain/repositories/collection.repository";
 
 @Injectable()
@@ -8,6 +8,12 @@ export class DeleteCollectionUseCase {
   ) {}
 
   async execute(id: string): Promise<void> {
+    const collection = await this.collectionRepository.findById(id);
+
+    if (!collection) {
+      throw new NotFoundException(`Collection with id ${id} not found`);
+    }
+
     return await this.collectionRepository.delete(id);
   }
 }
